Update the meta_plugin module with logging classes
The current logging method is not thread safe since logging will automatically add a new logging object to the UI and modify blender's context regardless of the thread.

Modifying blender's context from a thread other than the main thread is generally a bad idea and can cause issues. These new classes define a way of logging in a thread safe manner utilising a queue which only appends log items to the ui when flushed (if in muli-threaded mode).