Support blender 2.81a
Correctly select the right api product code based on bpy.app.version.

This is a very manual way of doing this and fairly messy. However there is already a much better way of handling different product versions in the dynamic-schema feature branch so this will do for now until that gets merged.