Remove upload and pack project file option
This option overlaps with the upload packed project option to such a degree that I think it adds more confusion than it saves.

I think it's better to let the user pack their project manually and then use the upload packed project option if they want to go this route.