Fix job preset option drop down
There is a bug in blender where if the value of an enum option is a string which python has lost it's reference too then the option does not display properly.

This commit changes the job options to be identified by the job preset id which is a string instead of creating a new string based on the job presets index (which gets lost).