Update render instaces input
Render instances should not default to the users job throughput limit but instead default to zero. This adds a checkbox to the render instaces input field so the user can select if they want to use the default value (0) or a specific value.

The max value for the integer input has also been removed as max_throughput might be wrong if it has been cached before the user changes their service level.